chore: run commitlint on pr title (#154)
Makes it easier to use consistent commit messages by checking PR
title, which can easily be fix by either the author or a core
developer. When we do the squash commit, this end-up as commit
message, one that will be commitlint compliant.

This is also making pre-commit.ci auto-updates compatible with
commitlint requirements.