Upgrade ansible-language-server to 0.4.0 (#344)
We also remove the README warning as 0.4.0 is fixing that bug.

We fix the bug in npm-sanity which picked any newer ALS version
instead of picking the last stable release. Using `ncu` tool instead
ensures only desirable versions are picked.

That change cannot be split due to the circular dependency between
pre-commit bug and the need to update ALS.