e2e-test: add hover tests (#370)
* add dummy collection to serve the purpose of testing extension over collections

* add hover test script and fixtures

* remove disable-extension argument while laumching vscode for test

The arguments are not necessary since the implementation of isolated execution of
tests using a user-dir and an extn-dir. The extn-dir has the listed extensions to be
used and only those extensions are enabled while the test runs in the isolated
environment.

* remove commented code

* fix typo

* fix typos in copy-paste leftovers in dummy collection

* make dummy collection name consistent throughout the code base

* fix failure because of ANSIBLE_FORCE_COLOR (fixes #373)

* Revert "remove disable-extension argument while launching vscode for test"

This reverts commit 163334c431c9cb6d3b5877f726f30b6a824f7b8d.
This also solves one more issue, that is, some unrelated formating changes
were there in this commit.