Enhance fallback to --syntax-check logic (#243)
* add patch for falling back to --syntax-check in all failure scenarios

* declutter try-catch block

* add logic to handle empty array scenario returned by executing ansible-lint command

* update code for progressTracker to avoif if stamements

* remove unnecessary falling back to syntax-check

After debugging, I found that when ansible-lint internally calls --synatax-check
and returns result, it is returned as an error. So, this had to be tackled in
order to avoid unncecessary fallback to --syntax-check during validation.

* remove unncessary console.log statements

* update the fallback logic

It has been found that ansible-lint is inconsistent in sending the
linting problems, as in sometimes it sends as error and sometimes
as simple result. This commit handles the cases and prevents
fallback to syntax-check in unnecessary scenarios.

* change warning message to info message in case there is nothing reported by ansible-lint

* remove 'no error' message from console

* add change note for the PR

* update the change note

* add spellings to c-spell dictionary

* update change note

* sort the dictionary