lightspeed: use the code value from the 403
The 403 error can now return a `code` value to specify the nature of the
error.
`permission_denied_wca_api_key_is_missing` is the first case we can get.
The error happens when the access is denied because the user has a seat
but the organization has not been fully configured yet.