CI fixes (#1026)
* GHA: Cache the ~/.cache/pre-commit/ directory

* GH/ci: fails on test error

* testRunner: expand the logic used to find the platform name

Prettier 3 and 4 are not formatting the block the same way. This change
simpifies the logic and avoid any trouble during the migration to prettier 4.

* GHA: CI ignore error on MacOS

* GHA/test: setup is a dependency

With need to prepare the venv before running the tests.

* task: inline the venv activate and the command

The following syntax for the `cmds` fields doesn't work as expected and the
first line triggers a failure. The venv envvars are never properly initialized.

```
- >
  source "${VIRTUAL_ENV}/bin/activate" &&
  yarn run test-e2e
```