GHA: don't assume go-task spawn a subshell to run the commands
Time to time, `go-task` fails to run the following command and returns an
err 127 (command not found). It seems it doesn't spawn a sub-shell to run
the 2 commands and consequently file to find a command with the same name.
By explicitly requesting a sub-shell with `bash -c 'command'`, we avoid the
error.

```
cmds:
      - source "${VIRTUAL_ENV}/bin/activate" && yarn run test-e2e
```

We also now print the Python version from the venv. It's a good way to
double-check the venv is actually created during the `setup`.