avoid an exception during init if 'roles' is not a directory
When the `roles` directory of the local workspace is not a regular directory,
`readdirSync()` raises an exception errcode (`ENOTDIR`).
The error was not catched and this was preventing the initialization of the
extension.

The commit also move a couple of functions to be able to properly unit-test
the change. We cannot import from the unit-test any module that use the
`vscode` object.

Finally, the commit changes the `LightSpeedManager.resetContext()` method to
make it a synchronous, this because the method changes an inner state of the
class that is consumed by other method.

Closes: #1084