LightspeedExplorer view as a Webview and new button
Convert the "Ansible Lighspeed Login" panel from a TreeView to a WebView. The goal
is to be able to expose an addition button icon.

This commit also adds a new "Explain the current playbook" button to this view to
trigger a playbook explanation.
Convert the "Ansible Lighspeed Login" panel from a TreeView to a WebView. The goal
is to be able to expose an addition button icon.

This commit also adds a new "Explain the current playbook" button to this view to
trigger a playbook explanation.