Update creator webviews to reflect ansible-creator CLI changes (#1582)
Adds a check for the ansible-creator version, as ansible-creator v24.10.0 is required for ansible-creator's new CLI structure to work with ADT server. 

Changes any reference of "scm-org" to "namespace" and "scm-project" to "collection". 

Adds a UI test for the Create Ansible Project Page webview.

Adds the "semver" dependency to do version comparison for ansible-creator.