AAP-37422: Empty line at the top of the file break completion
When the first line of the document is empty, the line is removed from the `documentLines`
array because of the early `trim()` call.
As a consequence, the `promptLine` value is incorrect and not completion is never triggered.
This PR address the problem.

See: c5b208d0985228110864dde9774c85cc68b09fe6