AAP-41894: lightspeed/roleGen: Delay the collection check
Until now, during the first page of the wizard, we forced the users to select a collection.
If no collection was found, we rise an error and prevent them from going forward.

Considering that it is not always trivial to create a collection and the user may want to
give a quick try to the role generation, we want to delay this check.