chore: remove boilerplate from test names
This makes it easier to navigate the test reports by avoiding
repetition of text that does not bring any value. It is obvious that
any test does "test/check/verify" something. The way mocha does
compose the tested tests, makes this issue even more serious and the
resulting test identifiers can become a nightmare to real.

Despite the fact that test name are configured with it() and
describe(), it does NOT mean that they are "descriptions". We can
still use comments to add details, but we should aim to keep the
arguments passed to these as short as possible and unique, with
names that should be easy to copy and search inside the code to
find them.

https://app.codecov.io/gh/ansible/vscode-ansible/tests/main